extract deleteModal
